refactor: moved user validation from server to gateway
